refactor : 동아리 점수 입력값 int에서 float로 변경

+++ src/main/java/ddingdong/ddingdongBE/domain/club/entity/Club.java
@@ -110,7 +110,7 @@ public class Club extends BaseEntity {
         return LocalDateTime.parse(inputLocalDateTimeFormat, formatter);
     }
 
-    public int editScore(Score score) {
+    public float editScore(Score score) {
         this.score = score;
 
         return this.score.getValue();
